Make hashi handle a single page loading strategy with script tags wrapped in template tags.
