+++ packages/hashi/package.json
@@ -17,8 +17,6 @@
     "eslint-plugin-compat": "^3.0.0",
     "express": "^4.16.4",
     "http-server": "^0.11.1",
-    "jest-environment-hashi-integration": "0.0.1",
-    "jest-environment-jsdom-script": "0.0.1",
     "webpack": "4.6.0",
     "webpack-cli": "^3.1.2"
   },
