+++ packages/hashi/src/replaceScript.js
@@ -1,75 +1,119 @@
+export function getScripts() {
+  const $scripts = document.querySelectorAll('template[hashi-script="true"]');
+  return [].map.call($scripts, $template => {
+    const parentNode = $template.parentNode;
+    const node = $template.content.children[0];
+    parentNode.removeChild($template);
+    // Keep a record of the parent node for later
+    // reattaching.
+    node._parentNode = parentNode;
+    return node;
+  });
+}
+
+function generateSingleInvocationFn(callback) {
+  return function() {
+    if (!callback.called) {
+      callback.called = true;
+      callback();
+    }
+  };
+}
+
 /*
  * Modified from https://ghinda.net/article/script-tags/
  */
-
-// https://html.spec.whatwg.org/multipage/scripting.html
-export const runScriptTypes = [
-  'application/javascript',
-  'application/ecmascript',
-  'application/x-ecmascript',
-  'application/x-javascript',
-  'text/ecmascript',
-  'text/javascript',
-  'text/javascript1.0',
-  'text/javascript1.1',
-  'text/javascript1.2',
-  'text/javascript1.3',
-  'text/javascript1.4',
-  'text/javascript1.5',
-  'text/jscript',
-  'text/livescript',
-  'text/x-ecmascript',
-  'text/x-javascript',
-];
-
-export default function replaceScript($script, callback) {
-  if (!$script.loaded) {
-    $script.loaded = true;
-    const s = document.createElement('script');
-    s.type = 'text/javascript';
-    [].forEach.call($script.attributes, attribute => {
-      // Set src later when everything else
-      // has been set.
-      if (attribute.name !== 'src') {
-        s.setAttribute(attribute.name, attribute.value);
+export function replaceScript($script, callback) {
+  if ($script) {
+    if (!$script.loaded) {
+      $script.loaded = true;
+      if ($script.src) {
+        $script.onload = callback;
+        $script.onerror = callback;
       }
-    });
-    if ($script.src) {
-      const cb = () => {
-        // Clean up onload and onerror handlers
-        // after they have been triggered to avoid
-        // these being called again.
-        try {
-          delete s.onload;
-          delete s.onerror;
-        } catch (e) {} // eslint-disable-line no-empty
+
+      const parentNode = $script._parentNode;
+
+      parentNode.appendChild($script);
+
+      // run the callback immediately for inline scripts
+      if (!$script.src) {
         callback();
-      };
-      s.onload = cb;
-      s.onerror = cb;
-      s.src = $script.src;
-      s.async = false;
+      }
+    }
+  } else {
+    callback();
+  }
+}
+
+/*
+ * runs an array of async functions in sequential order
+ */
+export function seq(arr, index) {
+  // first call, without an index
+  if (typeof index === 'undefined') {
+    index = 0;
+  }
+
+  function callback() {
+    index++;
+    if (index < arr.length) {
+      seq(arr, index);
     } else {
-      s.innerHTML = $script.innerHTML;
-    }
-
-    const parentNode = $script.parentNode;
-
-    const typeAttr = $script.getAttribute('type');
-
-    // only run script tags without the type attribute
-    // or with a javascript mime attribute value
-    if (!typeAttr || runScriptTypes.indexOf(typeAttr) !== -1) {
-      // Remove the element so that we don't clutter the DOM
-      // with duplicates.
-      parentNode.removeChild($script);
-      // re-insert the script tag so it executes.
-      parentNode.appendChild(s);
-    }
-
-    // run the callback immediately for inline scripts
-    if (!$script.src) {
-      callback();
+      // If finished trigger the DOM Content Loaded event and the load event
+      // Normally DOMContentLoaded would be triggered before the load event
+      // but we have already waited for the DOM to be inserted here, and
+      // then executed all the scripts, so we have a slightly less efficient
+      // rendering path, but we can be sure that both are appropriate to
+      // trigger at this point.
+      const DOMContentLoadedEvent = document.createEvent('Event');
+      DOMContentLoadedEvent.initEvent('DOMContentLoaded', true, true);
+      const loadEvent = document.createEvent('Event');
+      loadEvent.initEvent('load', true, true);
+      document.dispatchEvent(DOMContentLoadedEvent);
+      document.dispatchEvent(loadEvent);
+      window.dispatchEvent(loadEvent);
+      const elements = document.querySelectorAll(':not(script)');
+      Array.prototype.forEach.call(elements, element => {
+        element.dispatchEvent(loadEvent);
+      });
     }
   }
+
+  if (arr.length) {
+    const fn = arr[index];
+
+    fn(callback);
+  } else {
+    callback();
+  }
+}
+
+export function setScripts($scripts) {
+  // First generate the callbacks for all script tags.
+  // This will create insertion operations for all script tags
+
+  const $nonDeferredScripts = [].filter.call($scripts, script => !script.defer);
+  const $deferredScripts = [].filter.call($scripts, script => script.defer);
+
+  const runList = $nonDeferredScripts.concat($deferredScripts).map(function($script) {
+    return function(callback) {
+      const cb = generateSingleInvocationFn(callback);
+      window.onerror = cb;
+      try {
+        replaceScript($script, cb);
+      } catch (e) {
+        cb();
+      }
+    };
+  });
+
+  // insert the script tags sequentially
+  // to preserve execution order
+  seq(runList);
+}
+
+export function executePage() {
+  const scripts = getScripts();
+  setScripts(scripts);
 }
